refactor: code review fixes for data and pipeline modules
Major improvements:

**Type Safety (f64 → Decimal)**
- Migrate all price types from f64 to Decimal throughout data and pipeline layers
- Ensures no floating-point precision issues in financial calculations
- Affects: binance, coinbase, price_source, executor, settler modules

**Bug Fixes**
- Fix Coinbase timestamp: use exchange timestamp (root["timestamp"]) instead of local time
- Add [WS] prefix to all Coinbase WebSocket log messages for consistency

**Code Quality**
- Remove dead code: rpc_url field and query_usdc_balance function from polymarket.rs
- Consolidate balance queries: use BalanceChecker for initial balance in main.rs
- Extract shared test helper: new test_helpers.rs module eliminates 3 duplicate d() functions

**Documentation**
- Improve decider.rs time boundary documentation (clarify >=3min vs <2min thresholds)

**Testing**
- Add 2 unit tests for compute_filled_shares edge cases
- Total tests: 35 → 37

Verification: cargo check ✓, cargo clippy ✓, cargo test (37/37) ✓

+++ src/pipeline/price_source.rs
@@ -14,10 +14,12 @@ use crate::config::PriceSourceType;
 use crate::data::binance::BinanceClient;
 use crate::data::coinbase::CoinbaseClient;
 
+use rust_decimal::Decimal;
+
 /// Uniform ticker update shared across all price source backends.
 #[derive(Debug, Clone, Copy)]
 struct TickerUpdate {
-    price: f64,
+    price: Decimal,
     timestamp_ms: i64,
 }
 
@@ -41,7 +43,7 @@ impl From<crate::data::coinbase::TickerUpdate> for TickerUpdate {
 
 #[derive(Debug, Clone, Copy)]
 pub(crate) struct PriceTick {
-    pub price: f64,
+    pub price: Decimal,
     pub timestamp_ms: i64,
 }
 
@@ -77,7 +79,7 @@ impl PriceSource {
     }
 
     #[inline]
-    pub(crate) async fn latest(&self) -> Option<f64> {
+    pub(crate) async fn latest(&self) -> Option<Decimal> {
         self.buffer.read().await.back().map(|t| t.price)
     }
